Refactored code to different files
