+++ src/functions.cpp
@@ -1,0 +1,78 @@
+
+#include "functions.h"
+
+
+std::string str_join( const std::vector<std::string>& elements, const char* const separator)
+{
+  switch (elements.size())
+  {
+    case 0:
+      return "";
+    case 1:
+      return *(elements.begin());
+    default:
+      std::ostringstream os; 
+      std::copy(elements.cbegin(), elements.cend() - 1, std::ostream_iterator<std::string>(os, separator));
+      os << *elements.crbegin();
+      return os.str();
+  }
+}
+
+// Accepts a vector, transforms and returns a quantization logical vector
+// This function aims for space efficiency of the expression vector
+Quantile lognormalcdf(std::vector<int> ids, const Rcpp::NumericVector& v, unsigned int bits)
+{
+  Quantile expr;
+  expr.mu = std::accumulate(ids.begin(),ids.end(), 0, [&v](const double& mean, const int& index){
+      return  mean + v[index - 1];
+    }) / ids.size();
+  
+  expr.sigma = sqrt(std::accumulate(ids.begin(), ids.end(), 0, [&v, &expr](const double& variance, const int& index){
+        return pow(expr.mu - v[index - 1], 2);
+      }) / ids.size());
+  // initialize vector with zeros
+  expr.quantile.resize(ids.size() * bits, 0);
+  //std::cerr << "Mean,std" << expr.mu << "," << expr.sigma << std::endl;
+  //std::cerr << "ids size " << ids.size() << " v size " << v.size() << std::endl;
+  int expr_quantile_i = 0;
+  for (auto const& s : ids)
+  {
+    unsigned int t = round(normalCDF(v[s], expr.mu, expr.sigma) * (1 << bits));  
+    std::bitset<BITS> q = int2bin_core(t);
+    for (int i = 0; i < bits; ++i)
+    {
+       expr.quantile[expr_quantile_i++] = q[i];
+      
+    }
+  }
+  return expr;
+ }
+
+
+
+int byteToBoolVector(const std::vector<char> buf, std::vector<bool>& bool_vec)
+{
+  bool_vec.resize((buf.size() << 3), 0);
+  int c = 0;
+  for (auto const& b : buf)
+  {
+    for ( int i = 0; i < 8; i++)
+    {
+      bool_vec[c++] = ((b >> i) & 1);
+    }
+  }
+    
+}
+
+
+int getSizeBoolVector(const std::vector<bool>& v)
+{
+  int size = v.size() / 8;
+  if (v.size() % 8 != 0)
+  {
+    ++size;
+  }
+  return size;
+}
+
+
